Trait pallet_executor_registry::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber>;
    type StakeWeight: Parameter + Member + AtLeast32BitUnsigned + Codec + Default + Copy + MaybeSerializeDeserialize + Debug + MaxEncodedLen + TypeInfo + FixedPointOperand + From<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type MinExecutorStake: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type MaxExecutorStake: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type MinExecutors: Get<u32>;
    type MaxExecutors: Get<u32>;
    type MaxWithdrawals: Get<u32>;
    type WithdrawalDuration: Get<Self::BlockNumber>;
    type EpochDuration: Get<Self::BlockNumber>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber>

type StakeWeight: Parameter + Member + AtLeast32BitUnsigned + Codec + Default + Copy + MaybeSerializeDeserialize + Debug + MaxEncodedLen + TypeInfo + FixedPointOperand + From<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

The stake weight of an executor.

type MinExecutorStake: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

Minimum SSC required to be an executor.

type MaxExecutorStake: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

Maximum SSC that can be staked by a single executor.

type MinExecutors: Get<u32>

Minimum number of executors.

The minimum number of active executors is also constrained by this parameter with MIN_ACTIVE_EXECUTORS_FACTOR.

type MaxExecutors: Get<u32>

Maximum number of executors.

Increase this number gradually as the network grows.

type MaxWithdrawals: Get<u32>

Maximum number of ongoing unlocking items per executor.

type WithdrawalDuration: Get<Self::BlockNumber>

Number of blocks the withdrawn stake has to remain locked before it can become free.

Typically should be the same with fraud proof challenge period, like one week for arbitrum.

TODO: Use Slot instead of BlockNumber, which is closer to the actual elapsed time.

type EpochDuration: Get<Self::BlockNumber>

The amount of time each epoch should last in blocks.

The executor set for the bundle election is scheduled to rotate on each new epoch.